Refactor commonly used types to shared folder
