+++ src/handlers/platforms/next/create-app-register-handler.ts
@@ -8,16 +8,12 @@ import { SALEOR_API_URL_HEADER } from "@/const";
 import { createDebug } from "@/debug";
 import { fetchRemoteJwks } from "@/fetch-remote-jwks";
 import { getAppId } from "@/get-app-id";
+import { HookCallbackErrorParams } from "@/handlers/shared/create-app-register-handler-types";
 import { validateAllowSaleorUrls } from "@/handlers/shared/validate-allow-saleor-urls";
 import { withAuthTokenRequired } from "@/middleware";
 import { HasAPL } from "@/saleor-app";
 
 const debug = createDebug("createAppRegisterHandler");
-
-type HookCallbackErrorParams = {
-  status?: number;
-  message?: string;
-};
 
 class RegisterCallbackError extends Error {
   public status = 500;
