+++ src/handlers/platforms/next/saleor-webhooks/process-saleor-webhook.ts
@@ -3,52 +3,15 @@ import { NextApiRequest } from "next";
 import getRawBody from "raw-body";
 
 import { APL } from "@/APL";
-import { AuthData } from "@/APL/apl";
 import { createDebug } from "@/debug";
 import { fetchRemoteJwks } from "@/fetch-remote-jwks";
+import { WebhookContext, WebhookError } from "@/handlers/shared/process-saleor-webhook";
 import { getBaseUrl, getSaleorHeaders } from "@/headers";
 import { getOtelTracer } from "@/open-telemetry";
 import { parseSchemaVersion } from "@/util";
 import { verifySignatureWithJwks } from "@/verify-signature";
 
 const debug = createDebug("processSaleorWebhook");
-
-export type SaleorWebhookError =
-  | "OTHER"
-  | "MISSING_HOST_HEADER"
-  | "MISSING_DOMAIN_HEADER"
-  | "MISSING_API_URL_HEADER"
-  | "MISSING_EVENT_HEADER"
-  | "MISSING_PAYLOAD_HEADER"
-  | "MISSING_SIGNATURE_HEADER"
-  | "MISSING_REQUEST_BODY"
-  | "WRONG_EVENT"
-  | "NOT_REGISTERED"
-  | "SIGNATURE_VERIFICATION_FAILED"
-  | "WRONG_METHOD"
-  | "CANT_BE_PARSED"
-  | "CONFIGURATION_ERROR";
-
-export class WebhookError extends Error {
-  errorType: SaleorWebhookError = "OTHER";
-
-  constructor(message: string, errorType: SaleorWebhookError) {
-    super(message);
-    if (errorType) {
-      this.errorType = errorType;
-    }
-    Object.setPrototypeOf(this, WebhookError.prototype);
-  }
-}
-
-export type WebhookContext<T> = {
-  baseUrl: string;
-  event: string;
-  payload: T;
-  authData: AuthData;
-  /** For Saleor < 3.15 it will be null. */
-  schemaVersion: number | null;
-};
 
 interface ProcessSaleorWebhookArgs {
   req: NextApiRequest;
