+++ src/handlers/shared/protected-handler.ts
@@ -1,0 +1,20 @@
+export type SaleorProtectedHandlerError =
+  | "OTHER"
+  | "MISSING_HOST_HEADER"
+  | "MISSING_DOMAIN_HEADER"
+  | "MISSING_API_URL_HEADER"
+  | "MISSING_AUTHORIZATION_BEARER_HEADER"
+  | "NOT_REGISTERED"
+  | "JWT_VERIFICATION_FAILED"
+  | "NO_APP_ID";
+
+export const ProtectedHandlerErrorCodeMap: Record<SaleorProtectedHandlerError, number> = {
+  OTHER: 500,
+  MISSING_HOST_HEADER: 400,
+  MISSING_DOMAIN_HEADER: 400,
+  MISSING_API_URL_HEADER: 400,
+  NOT_REGISTERED: 401,
+  JWT_VERIFICATION_FAILED: 401,
+  NO_APP_ID: 401,
+  MISSING_AUTHORIZATION_BEARER_HEADER: 400,
+};
