+++ src/handlers/shared/saleor-webhook.ts
@@ -1,0 +1,19 @@
+import { ResultStatusCodes } from "./generic-adapter-use-case-types";
+import { SaleorWebhookError } from "./process-saleor-webhook";
+
+export const WebhookErrorCodeMap: Record<SaleorWebhookError, ResultStatusCodes> = {
+  OTHER: 500,
+  MISSING_HOST_HEADER: 400,
+  MISSING_DOMAIN_HEADER: 400,
+  MISSING_API_URL_HEADER: 400,
+  MISSING_EVENT_HEADER: 400,
+  MISSING_PAYLOAD_HEADER: 400,
+  MISSING_SIGNATURE_HEADER: 400,
+  MISSING_REQUEST_BODY: 400,
+  WRONG_EVENT: 400,
+  NOT_REGISTERED: 401,
+  SIGNATURE_VERIFICATION_FAILED: 401,
+  WRONG_METHOD: 405,
+  CANT_BE_PARSED: 400,
+  CONFIGURATION_ERROR: 500,
+};
